+++ packages/redux-connector/src/constants.ts
@@ -7,13 +7,22 @@
  *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-import { logger, createLogger } from './logger';
-import { healthcheck } from './express/healthcheck';
-import HTTPStatusCode from './express/types/HTTPStatusCode';
-import KafkaController from './kafka/KafkaController';
-import TopicUpdater from './kafka/TopicUpdater';
-import type { SchemaConfig, SubscriptionCallback, Headers } from './kafka/types';
 
-export { logger, createLogger, healthcheck, HTTPStatusCode, KafkaController, TopicUpdater };
+export const DEFAULT_EVENT_NAME = 'message';
+export const DEFAULT_CRITERIA = '@@socket/';
+export const DEFAULT_RESPONSE_ACTION_PREFIX = '@@socket/';
 
-export type { Headers, SchemaConfig, SubscriptionCallback };
+// Socket.io
+export const SIO_CONNECT_EVENT = 'connect';
+export const SIO_DISCONNECT_EVENT = 'disconnect';
+export const SIO_CONNECT_ERROR_EVENT = 'connect_error';
+export const SIO_SERVER_DISCONNECT_REASON = 'io server disconnect';
+
+// Incoming Redux Actions
+export const OPEN_SOCKET_CONNECTION = '@@reduxConnector/OPEN_SOCKET_CONNECTION';
+export const DISCONNECT_SOCKET_CONNECTION = '@@reduxConnector/DISCONNECT_SOCKET_CONNECTION';
+
+// Outgoing Redux Actions
+export const ON_SOCKET_CONNECT = '@@reduxConnector/ON_SOCKET_CONNECT';
+export const ON_SOCKET_DISCONNECT = '@@reduxConnector/ON_SOCKET_DISCONNECT';
+export const ON_SOCKET_CONNECTION_ERROR = '@@reduxConnector/ON_SOCKET_CONNECTION_ERROR';