+++ packages/socket-connector/src/services/Configuration/index.ts
@@ -7,13 +7,28 @@
  *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-import { logger, createLogger } from './logger';
-import { healthcheck } from './express/healthcheck';
-import HTTPStatusCode from './express/types/HTTPStatusCode';
-import KafkaController from './kafka/KafkaController';
-import TopicUpdater from './kafka/TopicUpdater';
-import type { SchemaConfig, SubscriptionCallback, Headers } from './kafka/types';
 
-export { logger, createLogger, healthcheck, HTTPStatusCode, KafkaController, TopicUpdater };
+import { MessageMocks } from '../../types/socketConnector';
 
-export type { Headers, SchemaConfig, SubscriptionCallback };
+class Configuration {
+  private static instance: Configuration;
+
+  private messageMocks: MessageMocks | undefined;
+
+  public static getInstance(): Configuration {
+    if (!this.instance) {
+      this.instance = new Configuration();
+    }
+    return this.instance;
+  }
+
+  public setMocks(messageMocks?: MessageMocks): void {
+    this.messageMocks = messageMocks;
+  }
+
+  public getMocks(): MessageMocks | undefined {
+    return this.messageMocks;
+  }
+}
+
+export default Configuration;